pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["S4toriX <nesterov.fortech@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.3"
djangorestframework = "^3.15.2"
django-environ = "^0.11.2"
gunicorn = "^23.0.0"
djangorestframework-simplejwt = "^5.3.1"
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
identify = "^2.6.3"
pre-commit = "^4.0.1"
orjson = "^3.10.12"
django-cors-headers = "^4.6.0"
minio = "^7.2.12"
mysqlclient = "^2.2.6"
ansible = "^11.1.0"
pydantic-settings = "^2.7.0"
kafka-python = "^2.0.2"
kafka = "^1.3.5"
six = "^1.17.0"
kafka-python-ng = "^2.2.3"
confluent-kafka = "^2.6.2"


[tool.poetry.group.dev.dependencies]
pytest-django = "^4.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
